[M]:使用新框架更新

+++ tinder-v3-scenario-api/src/main/java/com/juntai/tinder/entity/ForcesPlan.java
@@ -3,14 +3,14 @@ package com.juntai.tinder.entity;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-import java.io.Serializable;
-import java.time.LocalDateTime;
-
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.juntai.tinder.model.ForcesPlanBaseModel;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
 
 /**
  * <p>
@@ -52,9 +52,9 @@ public class ForcesPlan implements Serializable {
 
     private String taskType;
 
-    private Integer startTime;
+    private Long startTime;
 
-    private Integer endTime;
+    private Long endTime;
 
     /**
      * 计划动作
